feat(themes): add vscode light theme
